Complete Without Order Update

+++ app/Http/Controllers/api/v1/OrderController.php
@@ -3,12 +3,14 @@
 
 namespace App\Http\Controllers\api\v1;
 use App\Http\Controllers\api\v1\BaseController as BaseController;
+use App\Http\Requests\OrderRequest;
 use App\Models\Order;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 
 class OrderController extends BaseController
 {
+
+
     /**
      * Display a listing of the resource.
      *
@@ -16,6 +18,9 @@ class OrderController extends BaseController
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('order.view')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any Order !');
+        }
         $data['order'] =  Order::with('orderItems')->get();
         return $this->sendResponse($data, 'All Order List');
     }
@@ -36,8 +41,11 @@ class OrderController extends BaseController
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(OrderRequest $request)
     {
+        if (is_null($this->user) || !$this->user->can('order.store')) {
+            abort(403, 'Sorry !! You are Unauthorized to Store any Order !');
+        }
         $order =Order::create($request->only('customer_name','customer_mobile','address','district', 'total_price', 'user_id' ));
         return $this->sendResponse($order, 'Order Created Successfully');
     }
@@ -50,6 +58,9 @@ class OrderController extends BaseController
      */
     public function show(Order $order)
     {
+        if (is_null($this->user) || !$this->user->can('order.show')) {
+            abort(403, 'Sorry !! You are Unauthorized to show any Order !');
+        }
          $data['order'] = $order->with('orderItems')->first();
         return $this->sendResponse($data, 'All Order List');
     }
@@ -72,9 +83,11 @@ class OrderController extends BaseController
      * @param  \App\Models\Order  $order
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Order $order)
+    public function update(OrderRequest $request, Order $order)
     {
-        //
+        if (is_null($this->user) || !$this->user->can('order.update')) {
+            abort(403, 'Sorry !! You are Unauthorized to update any Order !');
+        }
     }
 
     /**
@@ -85,7 +98,20 @@ class OrderController extends BaseController
      */
     public function destroy(Order $order)
     {
+        if (is_null($this->user) || !$this->user->can('order.delete')) {
+            abort(403, 'Sorry !! You are Unauthorized to delete any Order !');
+        }
         $order->delete();
         return $this->sendResponse('Deleted', 'Order Deleted Successfully');
+    }
+
+    public function status(Order $order, Request $request)
+    {
+        if (is_null($this->user) || !$this->user->can('order.approve')) {
+            abort(403, 'Sorry !! You are Unauthorized to delete any Order !');
+        }
+
+        $order->update($request->only('status'));
+        return $this->sendResponse( $order, 'Order Deleted Successfully');
     }
 }
